requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
matplotlib>=3.7.0
PyYAML>=6.0
numpy>=1.24.0
Pillow>=10.0.0
xlrd>=2.0.1
